requirements.txt
python-dotenv
langchain
langchain-openai 
langchain-community
langchain-groq
streamlit
transformers    